refactor: 이미지 스와이퍼 컴포넌트 독립 폴더로 변경
