+++ src/components/PostCard/PostCard.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Link from 'next/link';
 import styled from 'styled-components';
-import { ImageSwiper } from './ImageSwiper';
+import { ImageSwiper } from '../ImageSwiper/ImageSwiper';
 
 const testImageObject = [
   {
